build 관련 오류 해결

+++ src/main/java/com/hwn/sw_project/dto/match/MatchedBenefitResponse.java
@@ -1,5 +1,6 @@
 package com.hwn.sw_project.dto.match;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record MatchedBenefitResponse(
@@ -7,6 +8,6 @@ public record MatchedBenefitResponse(
         Long userId,
         Long benefitId,
         String benefitTitle,
-        Double score,
+        BigDecimal score,
         Instant matchedAt
 ) {}
